long by int

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] price = {100,325,68769,3353,236878,1};
+        long[] price = {100,325,68769,3353,236878,1};
 
         SalesManager salesManager = new SalesManager(price);
